Added dryrun option

+++ src/functions/install_pkg/build_package.sh
@@ -2,22 +2,28 @@ build_package() {
     # Build package
     cd "${1}"
     source PKGBUILD
-    makedeb --convert --pkgname "${1}"
+    ( makedeb --convert --pkgname "${1}" ) | grep -v "Aborted by user. Exiting..."
+    wait
 
-    # Copy package to local repository
-    sudo rm /etc/mpm/repo/debs/"${1}"*.deb &> /dev/null || printf ""
-    sudo cp "${1}"_"$(version_gen)"_*.deb /etc/mpm/repo/debs/
+    # Run if the dryrun argument wasn't passed
+    if [[ "${arg_dryrun}" != "true" ]]; then
 
-    # Add package to database
-    local export_to_database=$(echo "${1}"\\"$(version_gen)"\\"${2}")
+        # Copy package to local repository
+        sudo rm /etc/mpm/repo/debs/"${1}"*.deb &> /dev/null || printf ""
+        sudo cp "${1}"_"$(version_gen)"_*.deb /etc/mpm/repo/debs/
 
-    # Runs if package exists in database - replaces package entry in database
-    if cat /etc/mpm/sources.db | awk -F '\' '{print $1}' | grep "${1}" &> /dev/null; then
-        sudo sed -i "s|${1}\\.*|${export_to_database}|" /etc/mpm/sources.db
+        # Add package to database
+        local export_to_database=$(echo "${1}"\\"$(version_gen)"\\"${2}")
 
-    # Runs every other time - adds package entry to database
-    else
-        echo "${export_to_database}" | sudo tee -a /etc/mpm/sources.db &> /dev/null
+        # Runs if package exists in database - replaces package entry in database
+        if cat /etc/mpm/sources.db | awk -F '\' '{print $1}' | grep "${1}" &> /dev/null; then
+            sudo sed -i "s|${1}\\.*|${export_to_database}|" /etc/mpm/sources.db
+
+            # Runs every other time - adds package entry to database
+        else
+            echo "${export_to_database}" | sudo tee -a /etc/mpm/sources.db &> /dev/null
+        fi
+
     fi
 
     cd ..
